refactor many things, now supports permissions

+++ api-v1/services/userService.js
@@ -8,7 +8,7 @@ const userService = {
 		var sql = `INSERT INTO users (user_id, user_password, email) VALUES (?,?,?)`;
 		var connection = toolKit.getConnection();
 
-		connection.query(sql, [user_id, password,   email], function(error, results, fields) {
+		connection.query(sql, [user_id, password, email], function(error, results, fields) {
 			if (error) throw error;
 			console.log('user created');
 			callback();
@@ -29,29 +29,44 @@ const userService = {
 	},
 
 	deleteUser(user_id, callback) {
+		if (user_id == 'admin') {
+			console.log('admin account cannot be deleted');
+		} else {
+			var sql = `DELETE FROM users WHERE user_id =?`;
+			var connection = toolKit.getConnection();
 
-		var sql = `DELETE FROM users WHERE user_id =?`;
-		var connection = toolKit.getConnection();
-
-		connection.query(sql, [user_id], function(error, results, fields) {
-			if (error) throw error;
-			console.log('user deleted');
-			callback();
-		});
-		connection.end();
+			connection.query(sql, [user_id], function(error, results, fields) {
+				if (error) throw error;
+				console.log('user deleted');
+				callback();
+			});
+			connection.end();
+		}
 	},
 
-	getUsers(callback){
+	getUsers(callback) {
 		var sql = `SELECT user_id, user_password, email FROM users`;
 		var connection = toolKit.getConnection();
 		connection.query(sql, [], function(error, results, fields) {
 			if (error) throw error;
 			console.log('getusers');
-			callback(results);
+			if (results) {
+				var list = [];
+				results.forEach((item, index) => {
+					list.push({
+						"user_id": item.user_id,
+						"password": "",
+						"email": item.email,
+						"authentication_token": ""
+					});
+				});
+				callback(list);
+			}
+			else callback();
 		});
 		connection.end();
 	},
-	
+
 	getUser(user_id, callback) {
 		var sql = `SELECT user_id, user_password, email FROM users WHERE user_id =?`;
 		var connection = toolKit.getConnection();
@@ -69,8 +84,8 @@ const userService = {
 		});
 		connection.end();
 	},
-	
+
 };
 
 
-module.exports = { userService};
+module.exports = { userService };
